feat(evolve): wire Digester/Planner/Seesaw into reflect pipeline (R7)
Integrate the new P1 modules into the reflect evolution loop (src/hooks/reflect.rs)
per the Architect's integration map (analyze -> digest -> build_proposals
-> curate -> gate+seesaw -> variants -> write):

- After analyze_session(): run digest_session() to build per-task digests;
  populate SessionAnalysis.persistent_failure[_categories] from them
- Build AdaptationLandscape from history + digests; emit an exploration
  hint when recommends_exploration() (unresolved persistent failure +
  untried structural edit types)
- After seeding: check the seesaw registry; if digests regress a solved
  task, skip further seeding and warn; update the registry with this round
- Clean up 4 clippy warnings (derive Default, rename from_str, drop
  redundant cast/map_or)

No behavior change for sessions without regressions. 602 lib tests green,
clippy clean.

Refs: gap-analysis ref-010 §3, Phase 1 integration

+++ src/evolve/digester.rs
@@ -276,7 +276,7 @@ fn parse_epoch(ts: &str) -> Option<i64> {
     // Days since Unix epoch (proleptic Gregorian, UTC). Formula from Howard Hinnant.
     let y = if mo <= 2 { y - 1 } else { y };
     let era = if y >= 0 { y } else { y - 399 } / 400;
-    let yoe = (y - era * 400) as i64; // [0, 399]
+    let yoe = y - era * 400; // [0, 399]
     let doy = (153 * (if mo > 2 { mo - 3 } else { mo + 9 }) + 2) / 5 + d - 1;
     let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy; // [0, 146096]
     let days = era * 146097 + doe - 719468;
